union_iterator: add an iterator-level union implementation
When both inputs are sorted but we don't own their storage, a
sorted merge is reasonable.

Modifications to the public interface
-------------------------------------

RangeVec::union
---------------

Now only accepts `IntoNormalizedRangeIter`s, in order to gather the
results in a single pass.

Simple wrapper around `NormalizedRangeIter::union`.

- Smoke tested

Additions to the public interface
=================================

method union of trait NormalizedRangeIter
-----------------------------------------

- Smoke tested
- Proptested against a bitmap oracle
- Proptested for idempotence and union with complement
- Proptested for idempotence and union with complement
- Fuzz tested to match the simpler (normalization-based) union
- Fuzz tested with empty set / universe