NormalizedRangeIter: add internal checks
When internal_checks are enabled, we compare the size_hint to the
actual result's size, to make sure the hint range is correct: while
iterators *are allowed* to return wildly incorrect size hints, we
expect our implementations to be correct, if sometimes loose.

We also double check that the return value is normalized.